requirements.txt
Django==1.11.3
pytz==2018.3
django-extensions==2.0.6
fcm-django==0.2.16
requests
djangorestframework
python-google-places==1.4.1
